Add files via upload
I created a login page. If the email ID and password are correct, the user was initially redirected to valid.html. However, I later changed this because I wanted to display all the values of the corresponding row on the page. To achieve this, I used another servlet, DisplayServlet, to display the data. I used a session to pass the data from one servlet to another.